+++ packages/rax-swiper/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.1.8
+
+- [fix] return instance ref in web
+- [fix] make slidePrev and slideNext same in web and miniapp
+
+## 0.1.7
+
+- [feat] support swiper in miniapp
+
+## 0.1.6
+
+- [fix] remove space in `swiper.css` `@font-face` url
+- [fix] type of `Swiper` & `SwiperSlide`
+
+## 0.1.5
+
+- [chore] update swiper dependency version to 6.8.4
+- [feat] keep the same with Swiper.js 6.8.4
+- [fix] event handler can't update
+
 ## 0.1.4
 
 - [chore] update swiper version to 6.7.1